+++ packages/common/tests/test-env.js
@@ -1,0 +1,12 @@
+const test = require('ava');
+
+const { getRequiredEnvVar } = require('../env');
+
+test('getRequiredEnvVar returns an environment value if defined', async (t) => {
+  const result = getRequiredEnvVar('testVar', { testVar: 'testvalue' });
+  t.is(result, 'testvalue');
+});
+
+test('getRequiredEnvVar throws error if not defined', async (t) => {
+  t.throws(() => getRequiredEnvVar('testVar', {}));
+});
