+++ packages/db/tests/connection.js
@@ -11,8 +11,7 @@ const fakeConnectionConfig = {
 test.before(async (t) => {
   const { knex } = proxyquire('../dist/connection.js', {
     './config': {
-      getConnectionConfigEnv: () => fakeConnectionConfig,
-      getSecretConnectionConfig: () => Promise.resolve(fakeConnectionConfig),
+      getConnectionConfig: () => Promise.resolve(fakeConnectionConfig),
     },
   });
   t.context.knex = knex;
@@ -21,10 +20,12 @@ test.before(async (t) => {
 test('knex returns expected Knex object with migration defined',
   async (t) => {
     const results = await t.context.knex({
-      migrationDir: 'testMigrationDir',
-      databaseCredentialSecretArn: 'randomSecret',
-      KNEX_ASYNC_STACK_TRACES: 'true',
-      KNEX_DEBUG: 'true',
+      env: {
+        migrationDir: 'testMigrationDir',
+        databaseCredentialSecretArn: 'randomSecret',
+        KNEX_ASYNC_STACK_TRACES: 'true',
+        KNEX_DEBUG: 'true',
+      },
     });
     t.is('testMigrationDir', results.migrate.config.directory);
     t.is('knex_migrations', results.migrate.config.tableName);
@@ -38,10 +39,12 @@ test('knex returns expected Knex object with migration defined',
 test('knex returns expected Knex object with optional config defined',
   async (t) => {
     const results = await t.context.knex({
-      migrationDir: 'testMigrationDir',
-      databaseCredentialSecretArn: 'randomSecret',
-      KNEX_DEBUG: 'true',
-      KNEX_ASYNC_STACK_TRACES: 'true',
+      env: {
+        migrationDir: 'testMigrationDir',
+        databaseCredentialSecretArn: 'randomSecret',
+        KNEX_DEBUG: 'true',
+        KNEX_ASYNC_STACK_TRACES: 'true',
+      },
     });
     t.is('testMigrationDir', results.migrate.config.directory);
     t.is('knex_migrations', results.migrate.config.tableName);
@@ -50,7 +53,9 @@ test('knex returns expected Knex object with optional config defined',
 test('knex returns Knex object with a default migration set when env.migrations is not defined',
   async (t) => {
     const results = await t.context.knex({
-      databaseCredentialSecretArn: 'randomSecret',
+      env: {
+        databaseCredentialSecretArn: 'randomSecret',
+      },
     });
     t.is('./migrations', results.migrate.config.directory);
     t.is('knex_migrations', results.migrate.config.tableName);
@@ -59,13 +64,15 @@ test('knex returns Knex object with a default migration set when env.migrations 
 test('knex returns expected Knex object with manual db configuraiton options set',
   async (t) => {
     const results = await t.context.knex({
-      migrationDir: 'testMigrationDir',
-      PG_HOST: 'localhost',
-      PG_USER: 'fakeUser',
-      PG_PASSWORD: 'fakePassword',
-      PG_DATABASE: 'fakeDb',
-      KNEX_ASYNC_STACK_TRACES: 'true',
-      KNEX_DEBUG: 'true',
+      env: {
+        migrationDir: 'testMigrationDir',
+        PG_HOST: 'localhost',
+        PG_USER: 'fakeUser',
+        PG_PASSWORD: 'fakePassword',
+        PG_DATABASE: 'fakeDb',
+        KNEX_ASYNC_STACK_TRACES: 'true',
+        KNEX_DEBUG: 'true',
+      },
     });
     t.is('testMigrationDir', results.migrate.config.directory);
     t.is('knex_migrations', results.migrate.config.tableName);
